requirements.txt
Flask==2.0.1
twilio==7.17.0
Pillow==9.5.0
requests==2.31.0
python-dotenv==1.0.0
gradio-client==0.5.0
gunicorn==20.1.0